Crate re_video

Video decoding library.

Modules

decode 🔒

Video frame decoding.

demux 🔒

Video demultiplexing.

h264 🔒

General H.264 utilities.

stable_index_deque 🔒

time 🔒

Structs

Chunk

One chunk of encoded video data, representing a single crate::SampleMetadata.

DecodeSettings

Settings for video decoding.

FFmpegVersion

A successfully parsed FFmpeg version.

Frame

One decoded video frame.

FrameContent

Data for a decoded frame on native targets.

FrameInfo

Meta information about a decoded video frame, as reported by the decoder.

GroupOfPictures

A Group of Pictures (GOP) always starts with an I(DR)-frame, followed by delta-frames.

SampleMetadata

A single sample in a video.

SamplesStatistics

Meta informationa about the video samples.

StableIndexDeque

A deque with an offset that allows indices into it to stay valid.

Time

A value in time units.

Timescale

The number of time units per second.

VideoDataDescription

Description of video data.

VideoEncodingDetails

Various information about how the video was encoded.

Enums

ChromaSubsamplingModes

Chroma subsampling mode.

DecodeError

DecodeHardwareAcceleration

How the video should be decoded.

DetectGopStartError

Failure reason for detect_gop_start.

FFmpegError

FFmpegVersionParseError

GopStartDetection

Result of a successful GOP detection.

PixelFormat

Pixel format/layout used by FrameContent::data.

TrackKind

VideoCodec

The basic codec family used to encode the video.

VideoLoadError

Errors that can occur when loading a video.

YuvMatrixCoefficients

Yuv matrix coefficients used by PixelFormat::Yuv.

YuvPixelLayout

Pixel layout used by PixelFormat::Yuv.

YuvRange

Yuv value range used by PixelFormat::Yuv.

Traits

AsyncDecoder

Interface for an asynchronous video decoder.

Functions

build_info

Returns information about this crate

detect_gop_start

Try to determine whether a frame chunk is the start of a GOP.

ffmpeg_download_url

Download URL for the latest version of FFmpeg on the current platform. None if the platform is not supported.

new_decoder

Creates a new async decoder for the given video data.

Type Aliases

DecodeResult

GopIndex

Index used for referencing into VideoDataDescription::gops.

SampleIndex

Index used for referencing into VideoDataDescription::samples.

TrackId