Struct re_video::demux::VideoData

pub struct VideoData {
    pub config: Config,
    pub timescale: Timescale,
    pub duration: Time,
    pub gops: Vec<GroupOfPictures>,
    pub samples: Vec<Sample>,
    pub data: Vec<u8>,
    pub mp4_tracks: BTreeMap<TrackId, Option<TrackKind>>,
}

Decoded video data.

Fields
config: Configtimescale: Timescale

How many time units are there per second.

duration: Time

Duration of the video, in time units.

gops: Vec<GroupOfPictures>

We split video into GOPs, each beginning with a key frame, followed by any number of delta frames.

samples: Vec<Sample>

Samples contain the byte offsets into data for each frame.

This list is sorted in ascending order of decode timestamps.

Samples must be decoded in decode-timestamp order, and should be presented in composition-timestamp order.

data: Vec<u8>

This array stores all data used by samples.

mp4_tracks: BTreeMap<TrackId, Option<TrackKind>>

All the tracks in the mp4; not just the video track.

Can be nice to show in a UI.

Implementations

impl VideoData

pub fn load_mp4(bytes: &[u8]) -> Result<Self, VideoLoadError>

impl VideoData

pub fn load_from_bytes( data: &[u8], media_type: &str ) -> Result<Self, VideoLoadError>

Loads a video from the given data.

TODO(andreas, jan): This should not copy the data, but instead store slices into a shared buffer. at the very least the should be a way to extract only metadata.

pub fn duration(&self) -> Duration

Length of the video.

pub fn width(&self) -> u32

Natural width of the video.

pub fn height(&self) -> u32

Natural height of the video.

pub fn human_readable_codec_string(&self) -> String

The codec used to encode the video.

pub fn num_samples(&self) -> usize

The number of samples in the video.

pub fn subsampling_mode(&self) -> Option<ChromaSubsamplingModes>

Returns the subsampling mode of the video.

Returns None if not detected or unknown.

pub fn bit_depth(&self) -> Option<u8>

Per color component bit depth.

Usually 8, but 10 for HDR (for example).

pub fn is_monochrome(&self) -> Option<bool>

Returns None if the mp4 doesn’t specify whether the video is monochrome or we haven’t yet implemented the logic to determine this.

pub fn frame_timestamps_ns(&self) -> impl Iterator<Item = i64> + '_

Determines the presentation timestamps of all frames inside a video, returning raw time values.

Returned timestamps are in nanoseconds since start and are guaranteed to be monotonically increasing.

pub fn get(&self, sample: &Sample) -> Option<Chunk>

Returns None if the sample is invalid/out-of-range.