re_video::decode::ffmpeg_h264
Module ffmpeg

Send video data to ffmpeg over CLI to decode it.

Structs

FFmpegCliH264Decoder

Decode H.264 video via ffmpeg over CLI

FFmpegFrameInfo 🔒

ffmpeg does not tell us the timestamp/duration of a given frame, so we need to remember it.

FFmpegProcessAndListener 🔒

Encapsulates the running of an ffmpeg process.

FrameBuffer 🔒

NaluStreamState 🔒

StdinWithShutdown 🔒

Wraps an stdin with a shared shutdown boolean.

Enums

Error

FFmpegFrameData 🔒

Functions

read_ffmpeg_output 🔒

sanitize_ffmpeg_log_message 🔒

Strips out buffer addresses from FFmpeg log messages so that we can use it with the log-once family of methods.

should_ignore_log_msg 🔒

Ignore some common output from ffmpeg.

write_avc_chunk_to_nalu_stream 🔒

write_bytes 🔒

write_ffmpeg_input 🔒